M1: real places - postcode in, playable PK3 out
Full pipeline against live data: postcodes.io -> BNG tile -> Environment
Agency 1m LIDAR over WCS + a single cached Overpass query -> UDMF -> PK3.

Terrain is quantised into absolute elevation bands (not per-tile min/max) so
neighbouring tiles agree along shared edges. Building heights come from
P90(DSM-DTM) inside each footprint, falling back to OSM tags when LIDAR gives
an implausible answer. Roads are buffered by tagged width and split per
contour band so they step with the ground.

Two constraints found by playing it, both now encoded:

- Doom will not step up more than 24 units (0.75m here), so the default
  contour step drops to 0.5m and contour_bands refuses anything larger.
  At 1m the terrain rendered perfectly and was completely unclimbable.
- On a two-sided line the lower texture is drawn on the side facing the
  LOWER sector, but the HIGHER sector is the thing standing up. Painting it
  from the wrong side gave every building the terrain's mud texture.

Also: contour bands are dilated past their simplification tolerance so
neighbours overlap rather than leaving hairline gaps that fell through to the
tile's lowest elevation and rendered as full-height cliffs. Sectors that end
up with no lines are dropped and renumbered, which silences GZDoom's load
warnings.

Verified: CT7 9RP (74 buildings, 1504 sectors) and a second Thanet tile (15 buildings,
878 sectors) both load in GZDoom 4.14.2 with zero console warnings.

+++ src/postcode2wad/demo.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 from shapely.geometry import Polygon, box
 
 from . import UNITS_PER_METRE as U
-from .geometry import SKY_FLAT, SectorSpec, Thing, build_geometry
+from .geometry import SectorSpec, Thing, build_geometry
 
 SKY_HEIGHT = 4096  # 128m of headroom; the sky flat means you never see it
 
@@ -28,7 +28,7 @@ TERRAIN_WALL = "ROCK4"
 
 def m(metres: float) -> int:
     """Metres -> map units, integer-snapped."""
-    return int(round(metres * U))
+    return round(metres * U)
 
 
 def build_m0_scene(size_m: float = 200.0):
